pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "marrow"
version = "0.1.0"
description = "Ambient intelligence that lives with you"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "elevenlabs>=1.9.0",
    "mss>=9.0.0",
    "sounddevice>=0.5.0",
    "faster-whisper>=1.0.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "pystray>=0.19.0",
    "aiofiles>=23.0.0",
    "numpy>=1.26.0",
    "psutil>=5.9.0",
    "uiautomation>=2.0.0",
    "browser-use>=0.12.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "python-docx>=1.0.0",
    "pypdf>=4.0.0",
    "keyboard>=0.13.0",
    "pyautogui>=0.9.54",
    "langchain-anthropic>=0.3.0",
    "firecrawl-py>=1.0.0",
    "openai>=1.40.0",
    "PyQt6>=6.6.0",
]

[project.scripts]
marrow = "marrow.main:main"

[tool.hatch.build.targets.wheel]
packages = ["marrow"]
